+++ spec/dataset_spec.rb
@@ -64,6 +64,10 @@ describe 'Dataset' do
       expect(@p.link).to be_an_instance_of(Array)
     end
 
+    it '#locale' do
+      expect(@p.locale).to be_an_instance_of(String)
+    end
+
     it '#metadata' do
       expect(@p.metadata).to be_an_instance_of(Hash)
     end