+++ spec/journal_spec.rb
@@ -36,6 +36,10 @@ describe 'Journal' do
       expect(@p.issn).to be_an_instance_of(String)
     end
 
+    it '#locale' do
+      expect(@p.locale).to be_an_instance_of(String)
+    end
+
     it '#metadata' do
       expect(@p.metadata).to be_an_instance_of(Hash)
     end