+++ spec/organisation_spec.rb
@@ -40,6 +40,10 @@ describe 'Organisation' do
       expect(@p.email).to be_an_instance_of(Array)
     end
 
+    it '#locale' do
+      expect(@p.locale).to be_an_instance_of(String)
+    end
+
     it '#metadata' do
       expect(@p.metadata).to be_an_instance_of(Hash)
     end