+++ spec/publication_spec.rb
@@ -52,6 +52,10 @@ describe 'Publication' do
       expect(@p.file).to be_an_instance_of(Array)
     end
 
+    it '#locale' do
+      expect(@p.locale).to be_an_instance_of(String)
+    end
+
     it '#metadata' do
       expect(@p.metadata).to be_an_instance_of(Hash)
     end