+++ spec/publisher_spec.rb
@@ -32,6 +32,10 @@ describe 'Publisher' do
       expect(@p.created).to be_an_instance_of(String)
     end
 
+    it '#locale' do
+      expect(@p.locale).to be_an_instance_of(String)
+    end
+
     it '#metadata' do
       expect(@p.metadata).to be_an_instance_of(Hash)
     end